fix(core): move review signature helpers out of the AI stack import path
Importing ai_reviewer.constants from github_app executed
ai_reviewer/__init__.py, which eagerly loads the full AIReviewer
client/tool stack for GitHub-only paths and roughly doubled import
time (Codex review P2 finding). The signature and branding helpers
now live in the dependency-free backend/core/branding.py so
lightweight importers stay lightweight; wording, idempotency and
behavior are unchanged.
